refactor(plugin): break the generator stuff out into it's own module/mixins
